backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.5
groq==0.25.0
semgrep==1.113.0
bandit==1.8.3
